+++ connectlib/__init__.py
@@ -1,8 +1,8 @@
 import logging
 
 from connectlib.__version__ import __version__  # noqa
-from connectlib.batch_indexer import NpIndexGenerator
 from connectlib.experiment import execute_experiment
+from connectlib.index_generator.np_index_generator import NpIndexGenerator
 from connectlib.logger import set_logging_level
 
 set_logging_level(loglevel=logging.INFO)