+++ connectlib/index_generator/__init__.py
@@ -1,0 +1,7 @@
+from connectlib.index_generator.base import BaseIndexGenerator
+from connectlib.index_generator.np_index_generator import NpIndexGenerator
+
+__all__ = [
+    "BaseIndexGenerator",
+    "NpIndexGenerator",
+]