+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- feat: make a base class for the index generator and document it (#85)
+
 ## [0.7.0] - 2022-02-01
 
 ### Fixed